Refactor ContentView into subviews
 Created CardView and BoardView
